requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4
# passlib 1.7.x expects bcrypt 3.x API
bcrypt>=3.2.0,<4.0.0
PyJWT>=2.9.0
python-dotenv>=1.0.1
pydantic-settings>=2.6.0
slowapi>=0.1.9
